+++ R/ML_StackedModel.R
@@ -18,7 +18,8 @@
 #' @return \code{StackedModel} class object that inherits from \code{MLModel}.
 #'
 #' @references
-#' Breiman, L. (1996) \emph{Stacked Regression.} Machine Learning, 24, 49--64.
+#' Breiman, L. (1996). Stacked regression. \emph{Machine Learning}, \emph{24},
+#' 49-64.
 #'
 #' @seealso \code{\link{fit}}, \code{\link{resample}}
 #'
@@ -43,7 +44,7 @@ StackedModel <- function(
 
   if (!is.null(weights)) stopifnot(length(weights) == length(base_learners))
 
-  slots <- combine_modelslots(base_learners, settings("response_types"))
+  slots <- combine_model_slots(base_learners, settings("response_types"))
   new("StackedModel",
     name = "StackedModel",
     label = "Stacked Regression",